Katta 配置修改，小 bug 修改，分词器采用 jieba，无关功能

+++ katta-core/src/main/java/com/ivyft/katta/lib/lucene/range/ShortRange.java
@@ -34,11 +34,11 @@ public class ShortRange extends Range {
         this.minInclusive = minInclusive;
         this.maxInclusive = maxInclusive;
 
-        if (!minInclusive && min != Long.MAX_VALUE) {
+        if (!minInclusive && min != Short.MAX_VALUE) {
             min++;
         }
 
-        if (!maxInclusive && max != Long.MIN_VALUE) {
+        if (!maxInclusive && max != Short.MIN_VALUE) {
             max--;
         }
